# artifacts/erp/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Al Ghani Wholesale Traders ERP API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: getUsers
      tags: [users]
      summary: List all users
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Create a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    User:
      type: object
      required: [id, name, email, role, isActive, createdAt, permissions]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [ceo, developer, manager, sales, accountant, warehouse, content]
        isActive:
          type: boolean
        permissions:
          type: array
          items:
            type: string
        createdAt:
          type: string
    UserInput:
      type: object
      required: [name, email, role, password]
      properties:
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [ceo, developer, manager, sales, accountant, warehouse, content]
        password:
          type: string
        permissions:
          type: array
          items:
            type: string
    UserUpdate:
      type: object
      properties:
        name:
          type: string
        role:
          type: string
          enum: [ceo, developer, manager, sales, accountant, warehouse, content]
        isActive:
          type: boolean
        permissions:
          type: array
          items:
            type: string
...